requirements.txt
torch==2.1.0
torchvision==0.16.0
fastapi==0.110.0
uvicorn==0.27.0
mlflow==2.11.0
pillow==10.0.0
pytest==7.4.0
python-multipart
